Adding a simple cluster usage based scaling policy. This will serve as an example to add more complex auto scaling policies in the future.

+++ pkg/scaler/types/types.go
@@ -1,6 +1,6 @@
 package types
 
 type Resource struct {
-	Cpu    uint32
+	Cpu    uint64
 	Memory uint64
 }
